requirements.txt
# PC操作支援アプリケーション - 必要なライブラリ
# Gemini (Google AI) 専用版

# AI API - Google Gemini
google-generativeai>=0.3.0

# 画像処理
Pillow>=9.0.0

# 入力監視
pynput>=1.7.0

# 音声認識
SpeechRecognition>=3.10.0
pyaudio>=0.2.11

# 音声出力
pygame>=2.5.0

# GUI (標準ライブラリのため不要だが、参考として記載)
# tkinter - Python標準ライブラリ